requirements.txt
fastapi
uvicorn
google-analytics-data
google-auth
openai
python-dotenv
requests
pandas>=1.5
